requirements.txt
python-telegram-bot==20.7
Jinja2==3.1.3
playwright==1.43.0
